% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_reconstruction_QP.R
\name{best_reconstruction_QP}
\alias{best_reconstruction_QP}
\title{Find "best" reconstruction of a target signature or spectrum from a set of signatures.}
\usage{
best_reconstruction_QP(
  target.sig,
  sig.universe,
  max.subset.size = NULL,
  method = "cosine",
  trim.less.than = 1e-10
)
}
\arguments{
\item{target.sig}{The signature to reconstruct.}

\item{sig.universe}{The universe of signatures from which to reconstruct
\code{target.sig}.}

\item{max.subset.size}{Maximum number of signatures to use to
reconstruct \code{target.sig}.}

\item{method}{As in \code{\link[philentropy]{dist_one_one}}, and used only
to find the final "best" reconstruction. The optimized exposures from which
to selected the "best" reconstruction are calculated using
\code{\link{optimize_exposure_QP}}, which uses \code{\link[quadprog]{solve.QP}}.}

\item{trim.less.than}{After optimizing exposures with
\code{\link{optimize_exposure_QP}}, remove exposures less than
\code{trim.less.than} and then re-optimize.}
}
\value{
A list .... TODO
}
\description{
Find "best" reconstruction of a target signature or spectrum from a set of signatures.
}
\details{
Should normally be fast if you do not specify \code{max.subset.size},
  but will be combinatorially slow if \code{max.subset.size} is large
  and \code{trim.less.than} is small or negative. So do not do that.
  TODO explain algorithm. TODO: tests
}
\examples{
set.seed(888)
sig.u <- do.call(cbind, lapply(1:6, function(x) runif(n = 96)))
rr <- best_reconstruction_QP(target.sig = sig.u[ , 1, drop = FALSE],
                            sig.universe = sig.u[ , 2:6])
names(rr)
rr$optimized.exposure
rr$similarity
rr <- best_reconstruction_QP(target.sig = sig.u[ , 1, drop = FALSE],
                            sig.universe = sig.u[ , 2:6],
                            max.subset.size = 3)
rr$optimized.exposure
rr$similarity
}
